services/core_api/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.37.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
httpx==0.28.1
pydantic-settings==2.11.0
PyJWT==2.10.1
email-validator==2.3.0
lyricsgenius==3.3.1
